+++ dio_cache_interceptor/lib/src/model/cache_strategy.dart
@@ -2,6 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
 import 'package:dio_cache_interceptor/src/util/contants.dart';
 import 'package:dio_cache_interceptor/src/util/http_date.dart';
+import 'package:dio_cache_interceptor/src/util/response_extension.dart';
 
 class CacheStrategy {
   final RequestOptions? request;
@@ -101,6 +102,9 @@ class CacheStrategyFactory {
     }
 
     if (response != null) {
+      // Skip download
+      if (response.isAttachment()) return false;
+
       final responseCaching = CacheControl.fromHeader(
         response.headers[cacheControlHeader],
       );